backend/requirements.txt
# FastAPI y servidor web
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Validación de datos
pydantic==2.11.7

# Variables de entorno
python-dotenv==1.1.1

# Cliente HTTP para pruebas
requests==2.32.4

# Dependencias futuras para ML
# scikit-learn==1.3.0
# spacy==3.7.0
# pandas==2.1.0
# numpy==1.24.0

# Base de datos (futuro)
# sqlalchemy==2.0.0
# alembic==1.12.0
